Fix the dylib on OSX.
As described in a new comment and the discussion on #1, the current gcc
package includes a dylib with a bad path that gets embedded in our
libpgplot.dylib. In this commit, we work around the problem. I also tidy up
the dependencies on libgfortran based on my improving understanding of how
these things all work. (The `gcc`, `libgcc`, and `libgfortran` packages all
provide overlapping sets of files, on OSX at least.)

We also get rid of some pkg-config invocations that don't work and aren't
needed.